services/brain/requirements.txt
# MQTT
paho-mqtt==1.6.1

# Async
asyncio==3.4.3
aiomqtt==1.0.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2
